Split up monitoring by subsystem.
